Android: Issue #158: PowerAuthAuthorizationHttpHeader moved to io.getlime.security.powerauth.sdk
